reorganize api, move main code to core package and expose to top level
