+++ docs/source/core.rst
@@ -1,0 +1,13 @@
+torchgeometry.core
+==================
+
+.. currentmodule:: torchgeometry.core
+
+.. toctree::
+   :maxdepth: 2
+
+   geometric
+   transformations
+   pinhole
+   conversions
+   warping
