+++ docs/source/index.rst
@@ -20,11 +20,7 @@ TGM focuses on Image and tensor warping functions such as:
    :maxdepth: 2
    :caption: Package Reference
 
-   geometric
-   transformations
-   pinhole
-   conversions
-   warping
+   core
    image
    losses
    contrib
